Implement new features
Added a lives management system which will check for a first run of the game, if it doesn't detect a lives folder, you generously will be given one extra life. To get more lives, you have to find a file you want to sacrifice, and place it in the folder that predates your death.

New jamming function, there's a 33% chance that the gun will jam either due to bad striking pin or FTF due to bad ammunition, only whenever the bullet is in the firing chamber.

New Hardcore features not accessible by main.py; in hardcore the jamming rate is %13.